Updated to the latest development version (r1956):
Reset random seed used for girl chat at the start of the game, should fix girl chat random to be constant in playbacks (not tested).
Use different name for the save directory for development version versus official release.
Fixed French character set "ç" to not to be stick with the previous character (resulted in no gap between "nç").